FIX: Final fix for types

+++ entity_linking/data/mention_entity_matching.py
@@ -17,8 +17,8 @@ class MentionEntityMatchingDataset(Dataset):
     def __getitem__(self, idx):
         item = {key: torch.tensor(val[idx]) for key, val in self.encodings.items()}
         item.update({
-            'mention_spans': torch.tensor(self.mention_spans[idx], dtype=torch.int),
-            'entity_indices': torch.tensor(self.entity_indices[idx], dtype=torch.int),
+            'mention_spans': torch.tensor(self.mention_spans[idx], dtype=torch.long),
+            'entity_indices': torch.tensor(self.entity_indices[idx], dtype=torch.long),
             'labels': torch.tensor(self.entity_labels[idx], dtype=torch.float)
         })
         return item
